requirements.txt
chumpy==0.70
easydict==1.13
h5py==3.10.0
imageio==2.34.1
imageio-ffmpeg==0.4.9
ipython==8.12.3
joblib==1.3.2
json-tricks==3.17.3
json5==0.9.22
jupyter==1.0.0
matplotlib==3.7.5
numpy==1.23.0
opencv-python==4.9.0.80
pandas==2.0.3
pycocotools==2.0.7
pyparsing==3.1.2
pyrender==0.1.45
PyYAML==6.0.1
scipy==1.10.1
smplx==0.1.28
tensorboardX==2.6.2.2
termcolor==2.4.0
tornado==6.4
tqdm==4.66.4
transforms3d==0.4.1
trimesh==4.1.7
zipp==3.19.0
